fix : 타이포 그라피 폰트 설정 수정
